Changes from review
