+++ droidlet/lowlevel/hello_robot/remote/remote_hello_realsense.py
@@ -16,7 +16,9 @@ import numpy as np
 import cv2
 import open3d as o3d
 from droidlet.lowlevel.hello_robot.remote.utils import transform_global_to_base, goto
+from droidlet.lowlevel.hello_robot.remote.lidar import Lidar
 from slam_pkg.utils import depth_util as du
+import obstacle_utils
 from obstacle_utils import is_obstacle
 from droidlet.dashboard.o3dviz import serialize as o3d_pickle
 from data_compression import *
@@ -40,6 +42,8 @@ class RemoteHelloRealsense(object):
 
     def __init__(self, bot):
         self.bot = bot
+        self._lidar = Lidar()
+        self._lidar.start()
         self._done = True
         self._connect_to_realsense()
         # Slam stuff
@@ -57,6 +61,17 @@ class RemoteHelloRealsense(object):
 
     def get_camera_transform(self):
         return self.bot.get_camera_transform()
+
+    def get_lidar_scan(self):
+        # returns tuple (timestamp, scan)
+        return self._lidar.get_latest_scan()
+
+    def is_lidar_obstacle(self):
+        lidar_scan = self._lidar.get_latest_scan()
+        if lidar_scan is None:
+            print("no scan")
+            return False
+        return obstacle_utils.is_lidar_obstacle(lidar_scan)
 
     def _connect_to_realsense(self):
         config = rs.config()
@@ -178,9 +193,9 @@ class RemoteHelloRealsense(object):
 
     def is_obstacle_in_front(self, return_viz=False):
         base_state = self.bot.get_base_state()
-        lidar_scan = self.bot.get_latest_lidar_scan()
+        lidar_scan = self.get_lidar_scan()
         pcd = self.get_open3d_pcd()
-        ret = is_obstacle(pcd, base_state, lidar=lidar_scan, max_dist=0.5, return_viz=return_viz)
+        ret = is_obstacle(pcd, base_state, lidar_scan=lidar_scan, max_dist=0.5, return_viz=return_viz)
         if return_viz:
             obstacle, cpcd, crop, bbox, rest = ret
             # cpcd = o3d_pickle(cpcd)
